rearrange code
